// app/src/services/enums.ts
export enum Gender {
  MALE = 'male',
  FEMALE = 'female',
}

export enum Body {
  SLIM = 'slim',
  AVERAGE = 'average',
  ATHLETIC = 'athletic',
  CURVY = 'curvy',
}

export enum Personality {
  INTROVERT = 'introvert',
  AMBIVERT = 'ambivert',
  EXTROVERT = 'extrovert',
}

export enum Goal {
  SERIOUS_RELATIONSHIP = 'serious_relationship',
  CASUAL_RELATIONSHIP = 'casual_relationship',
  OPEN_RELATIONSHIP = 'open_relationship',
  FRIENDSHIP = 'friendship',
  ACQUAINTANCESHIP = 'acquaintanceship',
}

export enum Activity {
  ART = 'art',
  BAR = 'bar',
  BBQ = 'bbq',
  BEACH = 'beach',
  BILLIARDS = 'billiards',
  BOWLING = 'bowling',
  CAFE = 'cafe',
  CANOEING = 'canoeing',
  CINEMA = 'cinema',
  CIRCUS = 'circus',
  CLIMBING = 'climbing',
  CONCERT = 'concert',
  CONFECTIONERY = 'confectionery',
  COOKING = 'cooking',
  CYCLING = 'cycling',
  DANCING = 'dancing',
  FISHING = 'fishing',
  FOOTBALL = 'football',
  GAMES = 'games',
  GYM = 'gym',
  HIKING = 'hiking',
  KARAOKE = 'karaoke',
  LUNAPARK = 'lunapark',
  MALL = 'mall',
  MUSEUM = 'museum',
  PARK = 'park',
  PARTY = 'party',
  PINGPONG = 'pingpong',
  PUB = 'pub',
  RESTAURANT = 'restaurant',
  RUNNING = 'running',
  SHOPPING = 'shopping',
  SKATING = 'skating',
  SNORKELING = 'snorkeling',
  SURFING = 'surfing',
  SWIMMING = 'swimming',
  TELEVISION = 'television',
  TENNIS = 'tennis',
  THEATER = 'theater',
  ZOO = 'zoo',
}